Update vc_prompt.sh
Setting `GITSTATUS` beforehand is a tad more expensive in case you're not currently in a repository but it avoids the unreliable repo detection that breaks when you have deeply nested directories. Also, alternate value expansion saves a few lines - I'm not sure zsh can deal with this but I think so.